feat: waitlist crud

+++ services/admin-api/src/blockchain/mechanics/waitlist/list/list.service.ts
@@ -1,20 +1,22 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
+import { ForbiddenException, Injectable, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Brackets, DeleteResult, FindOneOptions, FindOptionsWhere, Repository } from "typeorm";
 import { StandardMerkleTree } from "@openzeppelin/merkle-tree";
 
-import type { ISearchableDto, ISearchDto } from "@gemunion/types-collection";
+import type { ISearchDto } from "@gemunion/types-collection";
+import type { IWaitListListCreateDto, IWaitListListUpdateDto } from "@framework/types";
 
-import { IWaitListListCreateDto } from "./interfaces";
-import { WaitListListEntity } from "./list.entity";
-import { WaitListGenerateDto } from "./dto";
 import { UserEntity } from "../../../../infrastructure/user/user.entity";
+import { AssetService } from "../../../exchange/asset/asset.service";
+import { WaitListListEntity } from "./list.entity";
+import type { IWaitListGenerateDto } from "./interfaces";
 
 @Injectable()
 export class WaitListListService {
   constructor(
     @InjectRepository(WaitListListEntity)
     private readonly waitListListEntityRepository: Repository<WaitListListEntity>,
+    protected readonly assetService: AssetService,
   ) {}
 
   public async search(dto: Partial<ISearchDto>, userEntity: UserEntity): Promise<[Array<WaitListListEntity>, number]> {
@@ -62,10 +64,33 @@ export class WaitListListService {
     return this.waitListListEntityRepository.findOne({ where, ...options });
   }
 
+  public findOneWithRelations(where: FindOptionsWhere<WaitListListEntity>): Promise<WaitListListEntity | null> {
+    return this.findOne(where, {
+      join: {
+        alias: "waitlist",
+        leftJoinAndSelect: {
+          item: "waitlist.item",
+          item_components: "item.components",
+          item_template: "item_components.template",
+          item_contract: "item_components.contract",
+        },
+      },
+    });
+  }
+
   public async create(dto: IWaitListListCreateDto, userEntity: UserEntity): Promise<WaitListListEntity> {
+    const { item } = dto;
+
+    const assetEntity = await this.assetService.create({
+      components: [],
+    });
+
+    await this.assetService.update(assetEntity, item);
+
     return this.waitListListEntityRepository
       .create({
         ...dto,
+        item: assetEntity,
         merchantId: userEntity.merchantId,
       })
       .save();
@@ -73,15 +98,34 @@ export class WaitListListService {
 
   public async update(
     where: FindOptionsWhere<WaitListListEntity>,
-    dto: Partial<ISearchableDto>,
+    dto: IWaitListListUpdateDto,
+    userEntity: UserEntity,
   ): Promise<WaitListListEntity> {
-    const waitListListEntity = await this.findOne(where);
+    const { item, ...rest } = dto;
+
+    const waitListListEntity = await this.findOne(where, {
+      join: {
+        alias: "waitlist",
+        leftJoinAndSelect: {
+          item: "waitlist.item",
+          components: "item.components",
+        },
+      },
+    });
 
     if (!waitListListEntity) {
       throw new NotFoundException("waitlistNotFound");
     }
 
-    Object.assign(waitListListEntity, dto);
+    if (waitListListEntity.merchantId !== userEntity.merchantId) {
+      throw new ForbiddenException("insufficientPermissions");
+    }
+
+    Object.assign(waitListListEntity, rest);
+
+    if (item) {
+      await this.assetService.update(waitListListEntity.item, item);
+    }
 
     return waitListListEntity.save();
   }
@@ -95,11 +139,21 @@ export class WaitListListService {
     });
   }
 
-  public async delete(where: FindOptionsWhere<WaitListListEntity>): Promise<DeleteResult> {
+  public async delete(where: FindOptionsWhere<WaitListListEntity>, userEntity: UserEntity): Promise<DeleteResult> {
+    const waitListListEntity = await this.findOne(where);
+
+    if (!waitListListEntity) {
+      throw new NotFoundException("waitListListNotFound");
+    }
+
+    if (waitListListEntity.merchantId !== userEntity.merchantId) {
+      throw new ForbiddenException("insufficientPermissions");
+    }
+
     return this.waitListListEntityRepository.delete(where);
   }
 
-  public async generate(dto: WaitListGenerateDto): Promise<{ root: string }> {
+  public async generate(dto: IWaitListGenerateDto): Promise<{ root: string }> {
     const waitlistListEntity = await this.waitListListEntityRepository.findOne({
       where: { id: dto.listId },
       relations: { items: true },
